requirements.txt
asttokens==2.0.5
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
blinker==1.4
bs4==0.0.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.3.9
Corpus==0.4.2
cvxpy==1.2.1
cycler==0.11.0
Cython==0.29.30
decorator==5.1.1
distlib==0.3.3
dnspython==2.1.0
ecos==2.0.10
email-validator==1.1.3
executing==0.8.3
filelock==3.4.0
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.13.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fonttools==4.33.3
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
ipython==8.4.0
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.4.2
MarkupSafe==2.0.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nltk==3.7
numpy==1.22.4
osqp==0.6.2.post5
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.4.0
plotly==5.8.2
prompt-toolkit==3.0.29
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyportfolioopt==1.5.3
python-dateutil==2.8.2
pytz==2022.1
qdldl==0.1.5.post2
regex==2022.6.2
requests==2.28.0
scikit-learn==1.1.1
scikit-surprise==1.1.1
scipy==1.8.1
scs==3.2.0
seaborn==0.11.2
six==1.16.0
sklearn==0.0
Snowball==0.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.28
stack-data==0.3.0
stem==1.8.0
surprise==0.1
tenacity==8.0.1
threadpoolctl==3.1.0
tqdm==4.64.0
traitlets==5.3.0
urllib3==1.26.9
virtualenv==20.10.0
wcwidth==0.2.5
Werkzeug==2.0.2
wordnet==0.0.1b2
WTForms==3.0.0
